web/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
Flask
Flask-Babel
Flask-SocketIO
psycopg2
gunicorn
python-dateutil
networkx
scikit-learn
plotly
pandas
gevent
pytest
pytest-cov
leidenalg
python-igraph
redis
eventlet
python-dotenv
torch
sentence-transformers
google-api-python-client>=2.0.0
